[backend] feat : 리뷰 생성, 수정, 삭제 추가
[backend] feat : 리뷰 생성, 수정, 삭제 추가

+++ backend/yourwine/src/main/java/com/ssafy/yourwine/model/key/ReviewKey.java
@@ -15,5 +15,5 @@ public class ReviewKey implements Serializable{
 	private String userId;
 	
 	@Column(name = "wine_id")
-	private int wineId;
+	private Long wineId;
 }
